Tech Debt: Component-ize things
