+++ src/components/Header.svelte
@@ -1,0 +1,6 @@
+<h1 class="text-lg font-semibold text-cyan-50 uppercase text-center tracking-wider">
+	Tea Curious <span class="font-light">Tea Challenge</span>
+</h1>
+<h2 class="text-3xl sm:text-4xl font-bold text-cyan-50 text-center tracking-wider">
+	Moving Oolong!
+</h2>
